Split mkl proxy file & add bulk csc math funs
