+++ cla3p.mod/source/cla3p/bulk/csc_math.hpp
@@ -44,6 +44,59 @@ void gem_x_vec(op_t opA, uint_t m, uint_t n, T_Scalar alpha,
 		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
 		const T_Scalar *x, T_Scalar beta, T_Scalar *y);
 
+//
+// Update: y = beta * y + alpha * A * x
+//
+template <typename T_Scalar>
+void sym_x_vec(uplo_t uplo, uint_t n, T_Scalar alpha, 
+		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
+		const T_Scalar *x, T_Scalar beta, T_Scalar *y);
+
+//
+// Update: y = beta * y + alpha * A * x
+//
+template <typename T_Scalar>
+void hem_x_vec(uplo_t uplo, uint_t n, T_Scalar alpha, 
+		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
+		const T_Scalar *x, T_Scalar beta, T_Scalar *y);
+
+//
+// Update: C = beta * C + alpha * opA(A) * B
+// C(m x n)
+//
+template <typename T_Scalar>
+void gem_x_gem(op_t opA, uint_t m, uint_t n, uint_t k, T_Scalar alpha,
+		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
+		const T_Scalar *b, uint_t ldb, T_Scalar beta, T_Scalar *c, uint_t ldc);
+
+//
+// Update: C = beta * C + alpha * A * B
+// C(m x n)
+//
+template <typename T_Scalar>
+void sym_x_gem(uplo_t uplo, uint_t m, uint_t n, T_Scalar alpha,
+		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
+		const T_Scalar *b, uint_t ldb, T_Scalar beta, T_Scalar *c, uint_t ldc);
+
+//
+// Update: C = beta * C + alpha * A * B
+// C(m x n)
+//
+template <typename T_Scalar>
+void hem_x_gem(uplo_t uplo, uint_t m, uint_t n, T_Scalar alpha,
+		const int_t *colptr, const int_t *rowidx, const T_Scalar *values, 
+		const T_Scalar *b, uint_t ldb, T_Scalar beta, T_Scalar *c, uint_t ldc);
+
+//
+// Update: C = opA(A) * B
+// C(m x n)
+//
+template <typename T_Scalar>
+void gem_x_gem(op_t opA, uint_t m, uint_t n, uint_t k,
+		const int_t *colptrA, const int_t *rowidxA, const T_Scalar *valuesA, 
+		const int_t *colptrB, const int_t *rowidxB, const T_Scalar *valuesB, 
+		int_t **colptrC, int_t **rowidxC, T_Scalar **valuesC); 
+
 /*-------------------------------------------------*/
 } // namespace csc
 } // namespace bulk
