+++ cla3p.mod/source/cla3p/proxies/mkl_proxy.hpp
@@ -64,22 +64,6 @@ omatadd_macro(void, complex_t);
 omatadd_macro(void, complex8_t);
 #undef omatadd_macro
 
-template <typename T_Scalar>
-void csc_add(uint_t m, uint_t n, T_Scalar alpha,
-		const int_t* colptrA, const int_t* rowidxA, const T_Scalar* valuesA, 
-		const int_t* colptrB, const int_t* rowidxB, const T_Scalar* valuesB,
-		int_t **colptrC, int_t **rowidxC, T_Scalar **valuesC);
-
-template <typename T_Scalar>
-void csc_mv(prop_t propA, uplo_t uploA, uint_t m, uint_t n, T_Scalar alpha, op_t opA,
-		const int_t* colptrA, const int_t* rowidxA, const T_Scalar* valuesA, 
-		const T_Scalar* x, T_Scalar beta, T_Scalar *y);
-
-template <typename T_Scalar>
-void csc_mm(prop_t propA, uplo_t uploA, uint_t m, uint_t n, T_Scalar alpha, op_t opA,
-		const int_t* colptrA, const int_t* rowidxA, const T_Scalar* valuesA, 
-		uint_t k, const T_Scalar* b, uint_t ldb, T_Scalar beta, T_Scalar *c, uint_t ldc);
-
 /*-------------------------------------------------*/
 } // namespace mkl
 } // namespace cla3p
